LayoutTests/fast/forms/autocomplete-tokens-expected.txt

Tests for valid autocomplete tokens on input, select, and textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Valid keywords:
PASS input.autocomplete is "name"
PASS window.internals.autofillFieldName(input) is "name"
PASS input.autocomplete is "honorific-prefix"
PASS window.internals.autofillFieldName(input) is "honorific-prefix"
PASS input.autocomplete is "given-name"
PASS window.internals.autofillFieldName(input) is "given-name"
PASS input.autocomplete is "additional-name"
PASS window.internals.autofillFieldName(input) is "additional-name"
PASS input.autocomplete is "family-name"
PASS window.internals.autofillFieldName(input) is "family-name"
PASS input.autocomplete is "honorific-suffix"
PASS window.internals.autofillFieldName(input) is "honorific-suffix"
PASS input.autocomplete is "nickname"
PASS window.internals.autofillFieldName(input) is "nickname"
PASS input.autocomplete is "organization-title"
PASS window.internals.autofillFieldName(input) is "organization-title"
PASS input.autocomplete is "username"
PASS window.internals.autofillFieldName(input) is "username"
PASS input.autocomplete is "new-password"
PASS window.internals.autofillFieldName(input) is "new-password"
PASS input.autocomplete is "current-password"
PASS window.internals.autofillFieldName(input) is "current-password"
PASS input.autocomplete is "organization"
PASS window.internals.autofillFieldName(input) is "organization"
PASS input.autocomplete is "street-address"
PASS window.internals.autofillFieldName(input) is "street-address"
PASS input.autocomplete is "address-line1"
PASS window.internals.autofillFieldName(input) is "address-line1"
PASS input.autocomplete is "address-line2"
PASS window.internals.autofillFieldName(input) is "address-line2"
PASS input.autocomplete is "address-line3"
PASS window.internals.autofillFieldName(input) is "address-line3"
PASS input.autocomplete is "address-level4"
PASS window.internals.autofillFieldName(input) is "address-level4"
PASS input.autocomplete is "address-level3"
PASS window.internals.autofillFieldName(input) is "address-level3"
PASS input.autocomplete is "address-level2"
PASS window.internals.autofillFieldName(input) is "address-level2"
PASS input.autocomplete is "address-level1"
PASS window.internals.autofillFieldName(input) is "address-level1"
PASS input.autocomplete is "country"
PASS window.internals.autofillFieldName(input) is "country"
PASS input.autocomplete is "country-name"
PASS window.internals.autofillFieldName(input) is "country-name"
PASS input.autocomplete is "postal-code"
PASS window.internals.autofillFieldName(input) is "postal-code"
PASS input.autocomplete is "cc-name"
PASS window.internals.autofillFieldName(input) is "cc-name"
PASS input.autocomplete is "cc-given-name"
PASS window.internals.autofillFieldName(input) is "cc-given-name"
PASS input.autocomplete is "cc-additional-name"
PASS window.internals.autofillFieldName(input) is "cc-additional-name"
PASS input.autocomplete is "cc-family-name"
PASS window.internals.autofillFieldName(input) is "cc-family-name"
PASS input.autocomplete is "cc-number"
PASS window.internals.autofillFieldName(input) is "cc-number"
PASS input.autocomplete is "cc-exp"
PASS window.internals.autofillFieldName(input) is "cc-exp"
PASS input.autocomplete is "cc-exp-month"
PASS window.internals.autofillFieldName(input) is "cc-exp-month"
PASS input.autocomplete is "cc-exp-year"
PASS window.internals.autofillFieldName(input) is "cc-exp-year"
PASS input.autocomplete is "cc-csc"
PASS window.internals.autofillFieldName(input) is "cc-csc"
PASS input.autocomplete is "cc-type"
PASS window.internals.autofillFieldName(input) is "cc-type"
PASS input.autocomplete is "transaction-currency"
PASS window.internals.autofillFieldName(input) is "transaction-currency"
PASS input.autocomplete is "transaction-amount"
PASS window.internals.autofillFieldName(input) is "transaction-amount"
PASS input.autocomplete is "language"
PASS window.internals.autofillFieldName(input) is "language"
PASS input.autocomplete is "bday"
PASS window.internals.autofillFieldName(input) is "bday"
PASS input.autocomplete is "bday-day"
PASS window.internals.autofillFieldName(input) is "bday-day"
PASS input.autocomplete is "bday-month"
PASS window.internals.autofillFieldName(input) is "bday-month"
PASS input.autocomplete is "bday-year"
PASS window.internals.autofillFieldName(input) is "bday-year"
PASS input.autocomplete is "sex"
PASS window.internals.autofillFieldName(input) is "sex"
PASS input.autocomplete is "url"
PASS window.internals.autofillFieldName(input) is "url"
PASS input.autocomplete is "photo"
PASS window.internals.autofillFieldName(input) is "photo"
PASS input.autocomplete is "tel"
PASS window.internals.autofillFieldName(input) is "tel"
PASS input.autocomplete is "tel-country-code"
PASS window.internals.autofillFieldName(input) is "tel-country-code"
PASS input.autocomplete is "tel-national"
PASS window.internals.autofillFieldName(input) is "tel-national"
PASS input.autocomplete is "tel-area-code"
PASS window.internals.autofillFieldName(input) is "tel-area-code"
PASS input.autocomplete is "tel-local"
PASS window.internals.autofillFieldName(input) is "tel-local"
PASS input.autocomplete is "tel-local-prefix"
PASS window.internals.autofillFieldName(input) is "tel-local-prefix"
PASS input.autocomplete is "tel-local-suffix"
PASS window.internals.autofillFieldName(input) is "tel-local-suffix"
PASS input.autocomplete is "tel-extension"
PASS window.internals.autofillFieldName(input) is "tel-extension"
PASS input.autocomplete is "email"
PASS window.internals.autofillFieldName(input) is "email"
PASS input.autocomplete is "impp"
PASS window.internals.autofillFieldName(input) is "impp"

Contact scope:
PASS input.autocomplete is "home tel"
PASS window.internals.autofillFieldName(input) is "tel"
PASS input.autocomplete is "home tel-country-code"
PASS window.internals.autofillFieldName(input) is "tel-country-code"
PASS input.autocomplete is "home tel-national"
PASS window.internals.autofillFieldName(input) is "tel-national"
PASS input.autocomplete is "home tel-area-code"
PASS window.internals.autofillFieldName(input) is "tel-area-code"
PASS input.autocomplete is "home tel-local"
PASS window.internals.autofillFieldName(input) is "tel-local"
PASS input.autocomplete is "home tel-local-prefix"
PASS window.internals.autofillFieldName(input) is "tel-local-prefix"
PASS input.autocomplete is "home tel-local-suffix"
PASS window.internals.autofillFieldName(input) is "tel-local-suffix"
PASS input.autocomplete is "home tel-extension"
PASS window.internals.autofillFieldName(input) is "tel-extension"
PASS input.autocomplete is "home email"
PASS window.internals.autofillFieldName(input) is "email"
PASS input.autocomplete is "home impp"
PASS window.internals.autofillFieldName(input) is "impp"
PASS input.autocomplete is "work tel"
PASS window.internals.autofillFieldName(input) is "tel"
PASS input.autocomplete is "work tel-country-code"
PASS window.internals.autofillFieldName(input) is "tel-country-code"
PASS input.autocomplete is "work tel-national"
PASS window.internals.autofillFieldName(input) is "tel-national"
PASS input.autocomplete is "work tel-area-code"
PASS window.internals.autofillFieldName(input) is "tel-area-code"
PASS input.autocomplete is "work tel-local"
PASS window.internals.autofillFieldName(input) is "tel-local"
PASS input.autocomplete is "work tel-local-prefix"
PASS window.internals.autofillFieldName(input) is "tel-local-prefix"
PASS input.autocomplete is "work tel-local-suffix"
PASS window.internals.autofillFieldName(input) is "tel-local-suffix"
PASS input.autocomplete is "work tel-extension"
PASS window.internals.autofillFieldName(input) is "tel-extension"
PASS input.autocomplete is "work email"
PASS window.internals.autofillFieldName(input) is "email"
PASS input.autocomplete is "work impp"
PASS window.internals.autofillFieldName(input) is "impp"
PASS input.autocomplete is "mobile tel"
PASS window.internals.autofillFieldName(input) is "tel"
PASS input.autocomplete is "mobile tel-country-code"
PASS window.internals.autofillFieldName(input) is "tel-country-code"
PASS input.autocomplete is "mobile tel-national"
PASS window.internals.autofillFieldName(input) is "tel-national"
PASS input.autocomplete is "mobile tel-area-code"
PASS window.internals.autofillFieldName(input) is "tel-area-code"
PASS input.autocomplete is "mobile tel-local"
PASS window.internals.autofillFieldName(input) is "tel-local"
PASS input.autocomplete is "mobile tel-local-prefix"
PASS window.internals.autofillFieldName(input) is "tel-local-prefix"
PASS input.autocomplete is "mobile tel-local-suffix"
PASS window.internals.autofillFieldName(input) is "tel-local-suffix"
PASS input.autocomplete is "mobile tel-extension"
PASS window.internals.autofillFieldName(input) is "tel-extension"
PASS input.autocomplete is "mobile email"
PASS window.internals.autofillFieldName(input) is "email"
PASS input.autocomplete is "mobile impp"
PASS window.internals.autofillFieldName(input) is "impp"
PASS input.autocomplete is "fax tel"
PASS window.internals.autofillFieldName(input) is "tel"
PASS input.autocomplete is "fax tel-country-code"
PASS window.internals.autofillFieldName(input) is "tel-country-code"
PASS input.autocomplete is "fax tel-national"
PASS window.internals.autofillFieldName(input) is "tel-national"
PASS input.autocomplete is "fax tel-area-code"
PASS window.internals.autofillFieldName(input) is "tel-area-code"
PASS input.autocomplete is "fax tel-local"
PASS window.internals.autofillFieldName(input) is "tel-local"
PASS input.autocomplete is "fax tel-local-prefix"
PASS window.internals.autofillFieldName(input) is "tel-local-prefix"
PASS input.autocomplete is "fax tel-local-suffix"
PASS window.internals.autofillFieldName(input) is "tel-local-suffix"
PASS input.autocomplete is "fax tel-extension"
PASS window.internals.autofillFieldName(input) is "tel-extension"
PASS input.autocomplete is "fax email"
PASS window.internals.autofillFieldName(input) is "email"
PASS input.autocomplete is "fax impp"
PASS window.internals.autofillFieldName(input) is "impp"
PASS input.autocomplete is "pager tel"
PASS window.internals.autofillFieldName(input) is "tel"
PASS input.autocomplete is "pager tel-country-code"
PASS window.internals.autofillFieldName(input) is "tel-country-code"
PASS input.autocomplete is "pager tel-national"
PASS window.internals.autofillFieldName(input) is "tel-national"
PASS input.autocomplete is "pager tel-area-code"
PASS window.internals.autofillFieldName(input) is "tel-area-code"
PASS input.autocomplete is "pager tel-local"
PASS window.internals.autofillFieldName(input) is "tel-local"
PASS input.autocomplete is "pager tel-local-prefix"
PASS window.internals.autofillFieldName(input) is "tel-local-prefix"
PASS input.autocomplete is "pager tel-local-suffix"
PASS window.internals.autofillFieldName(input) is "tel-local-suffix"
PASS input.autocomplete is "pager tel-extension"
PASS window.internals.autofillFieldName(input) is "tel-extension"
PASS input.autocomplete is "pager email"
PASS window.internals.autofillFieldName(input) is "email"
PASS input.autocomplete is "pager impp"
PASS window.internals.autofillFieldName(input) is "impp"

Mode:
PASS input.autocomplete is "shipping street-address"
PASS window.internals.autofillFieldName(input) is "street-address"
PASS input.autocomplete is "shipping address-line1"
PASS window.internals.autofillFieldName(input) is "address-line1"
PASS input.autocomplete is "shipping address-line2"
PASS window.internals.autofillFieldName(input) is "address-line2"
PASS input.autocomplete is "shipping address-line3"
PASS window.internals.autofillFieldName(input) is "address-line3"
PASS input.autocomplete is "shipping address-level4"
PASS window.internals.autofillFieldName(input) is "address-level4"
PASS input.autocomplete is "shipping address-level3"
PASS window.internals.autofillFieldName(input) is "address-level3"
PASS input.autocomplete is "shipping address-level2"
PASS window.internals.autofillFieldName(input) is "address-level2"
PASS input.autocomplete is "shipping address-level1"
PASS window.internals.autofillFieldName(input) is "address-level1"
PASS input.autocomplete is "shipping country"
PASS window.internals.autofillFieldName(input) is "country"
PASS input.autocomplete is "shipping country-name"
PASS window.internals.autofillFieldName(input) is "country-name"
PASS input.autocomplete is "shipping postal-code"
PASS window.internals.autofillFieldName(input) is "postal-code"
PASS input.autocomplete is "billing street-address"
PASS window.internals.autofillFieldName(input) is "street-address"
PASS input.autocomplete is "billing address-line1"
PASS window.internals.autofillFieldName(input) is "address-line1"
PASS input.autocomplete is "billing address-line2"
PASS window.internals.autofillFieldName(input) is "address-line2"
PASS input.autocomplete is "billing address-line3"
PASS window.internals.autofillFieldName(input) is "address-line3"
PASS input.autocomplete is "billing address-level4"
PASS window.internals.autofillFieldName(input) is "address-level4"
PASS input.autocomplete is "billing address-level3"
PASS window.internals.autofillFieldName(input) is "address-level3"
PASS input.autocomplete is "billing address-level2"
PASS window.internals.autofillFieldName(input) is "address-level2"
PASS input.autocomplete is "billing address-level1"
PASS window.internals.autofillFieldName(input) is "address-level1"
PASS input.autocomplete is "billing country"
PASS window.internals.autofillFieldName(input) is "country"
PASS input.autocomplete is "billing country-name"
PASS window.internals.autofillFieldName(input) is "country-name"
PASS input.autocomplete is "billing postal-code"
PASS window.internals.autofillFieldName(input) is "postal-code"

Section:
PASS input.autocomplete is "section-parent name"
PASS window.internals.autofillFieldName(input) is "name"
PASS input.autocomplete is "section-parent honorific-prefix"
PASS window.internals.autofillFieldName(input) is "honorific-prefix"
PASS input.autocomplete is "section-parent given-name"
PASS window.internals.autofillFieldName(input) is "given-name"
PASS input.autocomplete is "section-parent additional-name"
PASS window.internals.autofillFieldName(input) is "additional-name"
PASS input.autocomplete is "section-parent family-name"
PASS window.internals.autofillFieldName(input) is "family-name"
PASS input.autocomplete is "section-parent honorific-suffix"
PASS window.internals.autofillFieldName(input) is "honorific-suffix"
PASS input.autocomplete is "section-parent nickname"
PASS window.internals.autofillFieldName(input) is "nickname"
PASS input.autocomplete is "section-parent organization-title"
PASS window.internals.autofillFieldName(input) is "organization-title"
PASS input.autocomplete is "section-parent username"
PASS window.internals.autofillFieldName(input) is "username"
PASS input.autocomplete is "section-parent new-password"
PASS window.internals.autofillFieldName(input) is "new-password"
PASS input.autocomplete is "section-parent current-password"
PASS window.internals.autofillFieldName(input) is "current-password"
PASS input.autocomplete is "section-parent organization"
PASS window.internals.autofillFieldName(input) is "organization"
PASS input.autocomplete is "section-parent street-address"
PASS window.internals.autofillFieldName(input) is "street-address"
PASS input.autocomplete is "section-parent address-line1"
PASS window.internals.autofillFieldName(input) is "address-line1"
PASS input.autocomplete is "section-parent address-line2"
PASS window.internals.autofillFieldName(input) is "address-line2"
PASS input.autocomplete is "section-parent address-line3"
PASS window.internals.autofillFieldName(input) is "address-line3"
PASS input.autocomplete is "section-parent address-level4"
PASS window.internals.autofillFieldName(input) is "address-level4"
PASS input.autocomplete is "section-parent address-level3"
PASS window.internals.autofillFieldName(input) is "address-level3"
PASS input.autocomplete is "section-parent address-level2"
PASS window.internals.autofillFieldName(input) is "address-level2"
PASS input.autocomplete is "section-parent address-level1"
PASS window.internals.autofillFieldName(input) is "address-level1"
PASS input.autocomplete is "section-parent country"
PASS window.internals.autofillFieldName(input) is "country"
PASS input.autocomplete is "section-parent country-name"
PASS window.internals.autofillFieldName(input) is "country-name"
PASS input.autocomplete is "section-parent postal-code"
PASS window.internals.autofillFieldName(input) is "postal-code"
PASS input.autocomplete is "section-parent cc-name"
PASS window.internals.autofillFieldName(input) is "cc-name"
PASS input.autocomplete is "section-parent cc-given-name"
PASS window.internals.autofillFieldName(input) is "cc-given-name"
PASS input.autocomplete is "section-parent cc-additional-name"
PASS window.internals.autofillFieldName(input) is "cc-additional-name"
PASS input.autocomplete is "section-parent cc-family-name"
PASS window.internals.autofillFieldName(input) is "cc-family-name"
PASS input.autocomplete is "section-parent cc-number"
PASS window.internals.autofillFieldName(input) is "cc-number"
PASS input.autocomplete is "section-parent cc-exp"
PASS window.internals.autofillFieldName(input) is "cc-exp"
PASS input.autocomplete is "section-parent cc-exp-month"
PASS window.internals.autofillFieldName(input) is "cc-exp-month"
PASS input.autocomplete is "section-parent cc-exp-year"
PASS window.internals.autofillFieldName(input) is "cc-exp-year"
PASS input.autocomplete is "section-parent cc-csc"
PASS window.internals.autofillFieldName(input) is "cc-csc"
PASS input.autocomplete is "section-parent cc-type"
PASS window.internals.autofillFieldName(input) is "cc-type"
PASS input.autocomplete is "section-parent transaction-currency"
PASS window.internals.autofillFieldName(input) is "transaction-currency"
PASS input.autocomplete is "section-parent transaction-amount"
PASS window.internals.autofillFieldName(input) is "transaction-amount"
PASS input.autocomplete is "section-parent language"
PASS window.internals.autofillFieldName(input) is "language"
PASS input.autocomplete is "section-parent bday"
PASS window.internals.autofillFieldName(input) is "bday"
PASS input.autocomplete is "section-parent bday-day"
PASS window.internals.autofillFieldName(input) is "bday-day"
PASS input.autocomplete is "section-parent bday-month"
PASS window.internals.autofillFieldName(input) is "bday-month"
PASS input.autocomplete is "section-parent bday-year"
PASS window.internals.autofillFieldName(input) is "bday-year"
PASS input.autocomplete is "section-parent sex"
PASS window.internals.autofillFieldName(input) is "sex"
PASS input.autocomplete is "section-parent url"
PASS window.internals.autofillFieldName(input) is "url"
PASS input.autocomplete is "section-parent photo"
PASS window.internals.autofillFieldName(input) is "photo"
PASS input.autocomplete is "section-parent tel"
PASS window.internals.autofillFieldName(input) is "tel"
PASS input.autocomplete is "section-parent tel-country-code"
PASS window.internals.autofillFieldName(input) is "tel-country-code"
PASS input.autocomplete is "section-parent tel-national"
PASS window.internals.autofillFieldName(input) is "tel-national"
PASS input.autocomplete is "section-parent tel-area-code"
PASS window.internals.autofillFieldName(input) is "tel-area-code"
PASS input.autocomplete is "section-parent tel-local"
PASS window.internals.autofillFieldName(input) is "tel-local"
PASS input.autocomplete is "section-parent tel-local-prefix"
PASS window.internals.autofillFieldName(input) is "tel-local-prefix"
PASS input.autocomplete is "section-parent tel-local-suffix"
PASS window.internals.autofillFieldName(input) is "tel-local-suffix"
PASS input.autocomplete is "section-parent tel-extension"
PASS window.internals.autofillFieldName(input) is "tel-extension"
PASS input.autocomplete is "section-parent email"
PASS window.internals.autofillFieldName(input) is "email"
PASS input.autocomplete is "section-parent impp"
PASS window.internals.autofillFieldName(input) is "impp"

Invalid combinations:
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS input.autocomplete is ""
PASS window.internals.autofillFieldName(input) is "on"
PASS inputInAutocompleteOffForm.autocomplete is ""
PASS window.internals.autofillFieldName(inputInAutocompleteOffForm) is "off"
PASS inputInAutocompleteOffForm.autocomplete is ""
PASS window.internals.autofillFieldName(inputInAutocompleteOffForm) is "on"

Non-autocompleteable controls:
PASS button.autocomplete is "invalid"
PASS keygen.autocomplete is "invalid"
PASS successfullyParsed is true

TEST COMPLETE
